requirements.txt
kafka-python==2.0.2
opencv-python==4.8.0.74
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0
pydantic==2.0.3
flask==3.0.2 